refactor: 사다리와 참여자관련 테스트 클래스 분리
